docs: rewrite Tag Tags and TagsPoplist stories
