+++ src/components/Tag/__stories__/Disabled.stories.tsx
@@ -1,0 +1,7 @@
+import { Template } from './Template.stories'
+
+export const Disabled = Template.bind({})
+
+Disabled.args = {
+  disabled: true,
+}
