+++ src/components/Tag/__stories__/Icons.stories.tsx
@@ -1,0 +1,16 @@
+import { Template } from './Template.stories'
+
+export const Icons = Template.bind({})
+
+Icons.parameters = {
+  docs: {
+    description: {
+      story:
+        'Add Icon on left side of your tag, You can define it using `icon` property.',
+    },
+  },
+}
+
+Icons.args = {
+  icon: 'check',
+}
