+++ src/components/Tag/__stories__/Template.stories.tsx
@@ -1,0 +1,6 @@
+import { ComponentStory } from '@storybook/react'
+import Tag from '..'
+
+export const Template: ComponentStory<typeof Tag> = args => (
+  <Tag {...args}>Tag</Tag>
+)
