+++ src/components/Tag/__stories__/index.stories.tsx
@@ -1,81 +1,14 @@
-import { Meta, Story } from '@storybook/react'
-import { ComponentProps } from 'react'
+import { Meta } from '@storybook/react'
 import Tag from '..'
-import { SENTIMENTS } from '../../../theme'
 
 export default {
   component: Tag,
   title: 'Components/Data Display/Tag',
 } as Meta
 
-const Template: Story<ComponentProps<typeof Tag>> = args => (
-  <Tag {...args}>Tag</Tag>
-)
-
-export const Default = Template.bind({})
-
-export const Variants = Template.bind({})
-Variants.parameters = {
-  docs: {
-    description: {
-      story:
-        'Variants defines different colors of you component. You can define it using `variant` property.',
-    },
-  },
-}
-Variants.decorators = [
-  () => (
-    <div style={{ display: 'flex', flexWrap: 'wrap', gap: 8 }}>
-      {SENTIMENTS.map(sentiment => (
-        <Tag key={sentiment} variant={sentiment} onClose={() => {}}>
-          {sentiment}
-        </Tag>
-      ))}
-    </div>
-  ),
-]
-
-export const Icons = Template.bind({})
-Icons.parameters = {
-  docs: {
-    description: {
-      story:
-        'Add Icon on left side of your tag, You can define it using `icon` property.',
-    },
-  },
-}
-Icons.args = {
-  icon: 'check',
-}
-
-export const onClose = Template.bind({})
-onClose.parameters = {
-  docs: {
-    description: {
-      story:
-        '`onClose` property to pass function trigger on click on close button.',
-    },
-  },
-}
-onClose.args = {
-  onClose: () => {},
-}
-
-export const Loading = Template.bind({})
-
-Loading.parameters = {
-  docs: {
-    description: {
-      story:
-        'Loading for async operation or other use cases, You can define it by passing `isLoading` property to true.',
-    },
-  },
-}
-Loading.args = {
-  isLoading: true,
-}
-
-export const Disabled = Template.bind({})
-Disabled.args = {
-  disabled: true,
-}
+export { Playground } from './Playground.stories'
+export { Variants } from './Variants.stories'
+export { Icons } from './Icons.stories'
+export { OnClose } from './OnClose.stories'
+export { Loading } from './Loading.stories'
+export { Disabled } from './Disabled.stories'
