+++ src/components/Tags/__stories__/Playground.stories.tsx
@@ -1,0 +1,7 @@
+import { Template } from './Template.stories'
+
+export const Playground = Template.bind({})
+
+Playground.args = {
+  tags: ['default'],
+}
