+++ src/components/Tags/__stories__/Template.stories.tsx
@@ -1,0 +1,4 @@
+import { ComponentStory } from '@storybook/react'
+import Tags from '..'
+
+export const Template: ComponentStory<typeof Tags> = args => <Tags {...args} />
