+++ src/components/TagsPoplist/__stories__/Template.stories.tsx
@@ -1,0 +1,6 @@
+import { ComponentStory } from '@storybook/react'
+import TagsPoplist from '..'
+
+export const Template: ComponentStory<typeof TagsPoplist> = args => (
+  <TagsPoplist {...args} />
+)
